feat: refactoring code and adding control on users

+++ src/RequestHandlers/questionsHandler.ts
@@ -8,7 +8,6 @@ import { get_correct_answers_count } from '../Helpers/answersHelper';
 import { get_game, persist_game_update } from '../Repositories/gamesRepository';
 import { GAMEID } from '../Validation/game';
 import { QuestionImportData, QUIZID } from '../Validation/quiz';
-import { get_quiz } from '../Repositories/quizzesRepository';
 import axios from 'axios';
 import he from 'he';
 import { delete_from_question, persist_option } from '../Repositories/optionsRepository';
@@ -19,76 +18,56 @@ export async function get_one(req: Request, res: Response) {
     const { game_id } = req.params;
 
     try {
-        assert(game_id, GAMEID);
-    } catch (error) {
-        res.status(400).json({ message: 'The game id is invalid' });
-        return;
-    }
+        const game = req.game;
 
-    try {
-        // Find quiz by id
-        const game = await get_game(game_id);
+        if (game) { // The game is not null cause of validation with checkGameAccess in gamesMiddleware, this condition is for ts errors
 
-        if (!game) {
-            return res.status(404).json({ error: 'Game not found' });
-        }
+            // Count the number of questions
+            const nb_questions_total = await get_total_questions_count(game.quizzesQuiz_id);
 
 
-        const user = req.user;
+            // Verify if the game is finished
+            if (game.current_question_index >= nb_questions_total) {
 
-        if (game.userUser_id) {
-            if (!user) {
-                return res.status(401).json({ error: 'Authentication required to access this game' });
+                res.status(200).json({
+                    game_finished: true,
+                    correct_answers_nb: await get_correct_answers_count(game_id),
+                    nb_questions_total: nb_questions_total,
+                    quiz_id: game.quizzesQuiz_id
+                });
+                return;
             }
-            if (game.userUser_id !== user.user_id) {
-                return res.status(403).json({ error: 'You are not authorized to access this game' });
+
+            // Find actual question
+            const question = await get_current_question(game.quizzesQuiz_id, game.current_question_index);
+
+            if (!question) {
+                res.status(404).json({ error: 'Question not found' });
+                return;
             }
-        }
+
+            // Find all options for questions
+            const options = question.options.map((option, index) => ({
+                option_index: option.option_index,
+                option_text: option.option_text
+            }));
 
 
-        // Count the number of questions
-        const nb_questions_total = await get_total_questions_count(game.quizzesQuiz_id);
 
-
-        // Verify if the game is finished
-        if (game.current_question_index >= nb_questions_total) {
-
-            return res.status(200).json({
-                game_finished: true,
-                correct_answers_nb: await get_correct_answers_count(game_id),
+            // Build the response
+            res.status(200).json({
+                game_finished: false,
+                question_text: question.question_text,
+                options: options,
+                question_index: question.question_index + 1,
                 nb_questions_total: nb_questions_total,
-                quiz_id: game.quizzesQuiz_id
+                correct_answers_nb: await get_correct_answers_count(game_id),
+                question_type: question.question_type,
+                question_difficulty: question.question_difficulty,
+                question_category: question.question_category,
+                quiz_id: game.quizzesQuiz_id,
             });
         }
-
-        // Find actual question
-        const question = await get_current_question(game.quizzesQuiz_id, game.current_question_index);
-
-        if (!question) {
-            return res.status(404).json({ error: 'Question not found' });
-        }
-
-        // Find all options for questions
-        const options = question.options.map((option, index) => ({
-            option_index: option.option_index,
-            option_text: option.option_text
-        }));
-
-
-
-        // Build the response
-        res.status(200).json({
-            game_finished: false,
-            question_text: question.question_text,
-            options: options,
-            question_index: question.question_index + 1,
-            nb_questions_total: nb_questions_total,
-            correct_answers_nb: await get_correct_answers_count(game_id),
-            question_type: question.question_type,
-            question_difficulty: question.question_difficulty,
-            question_category: question.question_category,
-            quiz_id: game.quizzesQuiz_id,
-        });
     } catch (error) {
         console.error(error);
         res.status(500).json({ error: 'Internal server error' });
@@ -98,12 +77,6 @@ export async function get_one(req: Request, res: Response) {
 
 export async function send_answer(req: Request, res: Response) {
     const { game_id } = req.params;
-    try {
-        assert(game_id, GAMEID);
-    } catch (error) {
-        res.status(400).json({ message: 'The game id is invalid' });
-        return;
-    }
 
 
     let { question_index, option_index } = req.body;
@@ -119,77 +92,68 @@ export async function send_answer(req: Request, res: Response) {
     question_index--;
 
     try {
-        // Find the game by his id
-        const game = await get_game(game_id);
+        const game = req.game;
 
-        if (!game) {
-            return res.status(404).json({ error: 'Game not found' });
-        }
-
-        const user = req.user;
-
-        if (game.userUser_id) {
-            if (!user) {
-                return res.status(401).json({ error: 'Authentication required to access this game' });
+        if (game) { // The game is not null cause of validation with checkGameAccess in gamesMiddleware, this condition is for ts errors
+            // Verify that there in no desynchronization
+            if (question_index !== game.current_question_index) {
+                res.status(400).json({ error: 'Question\'s index invalid' });
+                return;
             }
-            if (game.userUser_id !== user.user_id) {
-                return res.status(403).json({ error: 'You are not authorized to access this game' });
+
+            // Count the number of questions
+            const nb_questions_total = await get_total_questions_count(game.quizzesQuiz_id);
+
+            if (question_index >= nb_questions_total) {
+                res.status(400).json({ error: 'Quiz is finished' });
+                return;
             }
+
+            // Find corresponding question
+            const question = await get_current_question(game.quizzesQuiz_id, question_index);
+
+            if (!question) {
+                res.status(404).json({ error: 'Question not found' });
+                return;
+            }
+
+            // Find chosen option
+            const chosenOption = question.options.find(
+                (option) => option.option_index === option_index
+            );
+
+            if (!chosenOption) {
+                res.status(400).json({ error: 'Invalid option index' });
+                return;
+            }
+
+            // Determine if the answer is correct
+            const isCorrect = chosenOption.is_correct_answer;
+
+            // Find correct answer's index
+            const correctOption = question.options.find(
+                (option) => option.is_correct_answer
+            );
+
+            if (!correctOption) {
+                res.status(500).json({ error: 'Correct answer not found' });
+                return;
+            }
+
+            const correctOptionIndex = correctOption.option_index;
+
+            // Update game in DB
+            await persist_game_update(game_id, game.current_question_index + 1);
+
+            // Add answer
+            await persist_answer(game_id, question.question_id, chosenOption.option_id,);
+
+            // Build the response
+            res.status(200).json({
+                is_correct: isCorrect,
+                correct_option_index: correctOptionIndex,
+            });
         }
-
-        // Verify that there in no desynchronization
-        if (question_index !== game.current_question_index) {
-            return res.status(400).json({ error: 'Question\'s index invalid' });
-        }
-
-        // Count the number of questions
-        const nb_questions_total = await get_total_questions_count(game.quizzesQuiz_id);
-
-        if (question_index >= nb_questions_total) {
-            return res.status(400).json({ error: 'Quiz is finished' });
-        }
-
-        // Find corresponding question
-        const question = await get_current_question(game.quizzesQuiz_id, question_index);
-
-        if (!question) {
-            return res.status(404).json({ error: 'Question not found' });
-        }
-
-        // Find chosen option
-        const chosenOption = question.options.find(
-            (option) => option.option_index === option_index
-        );
-
-        if (!chosenOption) {
-            return res.status(400).json({ error: 'Invalid option index' });
-        }
-
-        // Determine if the answer is correct
-        const isCorrect = chosenOption.is_correct_answer;
-
-        // Find correct answer's index
-        const correctOption = question.options.find(
-            (option) => option.is_correct_answer
-        );
-
-        if (!correctOption) {
-            return res.status(500).json({ error: 'Correct answer not found' });
-        }
-
-        const correctOptionIndex = correctOption.option_index;
-
-        // Update game in DB
-        await persist_game_update(game_id, game.current_question_index + 1);
-
-        // Add answer
-        await persist_answer(game_id, question.question_id, chosenOption.option_id,);
-
-        // Build the response
-        res.status(200).json({
-            is_correct: isCorrect,
-            correct_option_index: correctOptionIndex,
-        });
     } catch (error) {
         console.error(error);
         res.status(500).json({ error: 'Internal server error' });
@@ -202,13 +166,6 @@ export async function import_questions(req: Request, res: Response) {
     const { quiz_id } = req.params;
 
     try {
-        assert(quiz_id, QUIZID);
-    } catch (error) {
-        res.status(400).json({ message: 'The quiz id is invalid' });
-        return;
-    }
-
-    try {
         assert(req.body, QuestionImportData);
     } catch (error) {
         res.status(400).json({ message: 'Data is invalid: \n- category must be a number between 9 and 32\n- difficulty must be a string\n- amount must be a number between 1 and 50' });
@@ -216,12 +173,6 @@ export async function import_questions(req: Request, res: Response) {
     }
 
     try {
-        const quiz = await get_quiz(quiz_id);
-
-        if (!quiz) {
-            return res.status(404).json({ error: 'Quiz not found' });
-        }
-
         const { category, difficulty, amount } = req.body;
 
         // Build object params
@@ -244,7 +195,8 @@ export async function import_questions(req: Request, res: Response) {
         const { response_code, results } = apiResponse.data;
 
         if (response_code !== 0) {
-            return res.status(400).json({ error: 'Error retrieving questions from the API.' });
+            res.status(400).json({ error: 'Error retrieving questions from the API.' });
+            return;
         }
 
         // Browsing questions and options
@@ -257,7 +209,7 @@ export async function import_questions(req: Request, res: Response) {
             const incorrectAnswers = questionData.incorrect_answers.map((ans: string) => he.decode(ans));
 
             // Create question
-            const question = await persist_question(index, questionText, questionData.category, questionData.difficulty, questionData.type, quiz.quiz_id);
+            const question = await persist_question(index, questionText, questionData.category, questionData.difficulty, questionData.type, req.params.quiz_id);
 
             // Prepare the options
             const optionsData = [];
@@ -305,24 +257,16 @@ export async function get_informations(req: Request, res: Response) {
     const { quiz_id, question_id } = req.params;
 
     try {
-        assert(quiz_id, QUIZID);
-    } catch (error) {
-        res.status(400).json({ message: 'The quiz id is invalid' });
-        return;
-    }
-
-    try {
-        assert(question_id, UUID);
-    } catch (error) {
-        res.status(400).json({ message: 'The question id is invalid' });
-        return;
-    }
-
-    try {
         const question = await get_question_informations(question_id);
 
         if (!question) {
-            return res.status(404).json({ error: 'Question not found' });
+            res.status(404).json({ error: 'Question not found' });
+            return;
+        }
+
+        if (question.quizzesQuiz_id !== quiz_id) {
+            res.status(404).json({ error: 'Question does not belong to the provided quiz' });
+            return;
         }
 
         res.status(201).json({ message: 'Question found', question });
@@ -338,13 +282,6 @@ export async function create_one(req: Request, res: Response) {
     const { quiz_id } = req.params;
 
     try {
-        assert(quiz_id, QUIZID);
-    } catch (error) {
-        res.status(400).json({ message: 'The quiz id is invalid' });
-        return;
-    }
-
-    try {
         assert(req.body, QuestionCreationData);
     } catch (error) {
         res.status(400).json({ message: 'Data is invalid' });
@@ -352,12 +289,6 @@ export async function create_one(req: Request, res: Response) {
     }
 
     try {
-        const quiz = await get_quiz(quiz_id);
-
-        if (!quiz) {
-            return res.status(404).json({ error: 'Quiz not found' });
-        }
-
         const {
             question_text,
             question_difficulty,
@@ -369,13 +300,14 @@ export async function create_one(req: Request, res: Response) {
         // Verify that there is a correct answer
         const correctOptions = options.filter((option) => option.is_correct_answer === true);
         if (correctOptions.length != 1) {
-            return res.status(400).json({ error: 'There must be a correct answer' });
+            res.status(400).json({ error: 'There must be a correct answer' });
+            return;
         }
 
 
         const question_index = await get_total_questions_count(quiz_id);
 
-        const question = await persist_question(question_index, question_text, question_category, question_difficulty, question_type, quiz.quiz_id);
+        const question = await persist_question(question_index, question_text, question_category, question_difficulty, question_type, req.params.quiz_id);
 
         const optionsData = options.map((option: any, index: number) => ({
             option_text: option.option_text,
@@ -410,13 +342,18 @@ export async function delete_one(req: Request, res: Response) {
     const { quiz_id, question_id } = req.params;
 
     try {
-        assert(quiz_id, QUIZID);
-    } catch (error) {
-        res.status(400).json({ message: 'The quiz id is invalid' });
-        return;
-    }
+        const question = await get_question_informations(question_id);
 
-    try {
+        if (!question) {
+            res.status(404).json({ error: 'Question not found' });
+            return;
+        }
+
+        if (question.quizzesQuiz_id !== quiz_id) {
+            res.status(404).json({ error: 'Question does not belong to the provided quiz' });
+            return;
+        }
+
         await delete_question(question_id);
 
         const remainingQuestions = await get_all_questions(quiz_id);
@@ -441,12 +378,6 @@ export async function update_one(req: Request, res: Response) {
     const { quiz_id, question_id } = req.params;
 
     try {
-        assert(quiz_id, QUIZID);
-    } catch (error) {
-        res.status(400).json({ message: 'The quiz id is invalid' });
-        return;
-    }
-    try {
         assert(req.body, QuestionUpdateData);
     } catch (error) {
         res.status(400).json({ message: 'Data is invalid' });
@@ -458,6 +389,11 @@ export async function update_one(req: Request, res: Response) {
 
         if (!existingQuestion) {
             res.status(404).json({ message: 'Question not found' });
+            return;
+        }
+
+        if (existingQuestion.quizzesQuiz_id !== quiz_id) {
+            res.status(404).json({ error: 'Question does not belong to the provided quiz' });
             return;
         }
